requirements.txt
Django
psycopg2-binary
django-environ
djangorestframework
djangorestframework-simplejwt
django-cors-headers
social-auth-core
social-auth-app-django